chore(crypto): CRP-2788 Rename IBE parameter from context to identity
For one this avoids confusion with the other value called context in the larger
VetKey protocol, which has nothing to do with this parameter.  Calling it
`identity` captures its meaning within the context of the IBE, and the doc
comments are ammended to explain the connection between the `identity` parameter
in IBE and the `input` parameter in VetKey derivation.