chore(EXC-2035): Upgrade Wasmtime to v.32 (#4927)
Upgrades to Wasmtime v.32.

The main changes (besides bumping versions) is in the custom patches
that we maintain related to determinism and cache issues with bazel
builds. More specifically, the changes are:

* Wasmtime 32 contains a
[refactor](https://github.com/bytecodealliance/wasmtime/pull/10352)
which changes the isle code generation such that the following patches
are no longer necessary and were removed:
  * `bazel/cranelift-assembler-main.patch`
  * `bazel/cranelift-codegen-meta-isle.patch`
* `bazel/cranelift-assembler-lib.patch` is still needed but it’s changed
to return an empty vector instead of including the path to generated
files as it’s only used in `src/main.rs` which we never call during the
build process.
* Wasmtime 32 also contains another
[refactor](https://github.com/bytecodealliance/wasmtime/commit/c3bf042acb9999962155ae07f54015ff835a8b92)
that moves `srcgen` to its own crate, which requires updating
`bazel/cranelift-codegen-meta.patch`.
* bazel/cranelift-isle.patch` was updated to remove one more case where
a path was printed to a generated file.

---------

Co-authored-by: Adam Bratschi-Kaye <adam.bratschikaye@dfinity.org>
Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>