feat: CRP-2618 Allow zero `pre_signatures_to_create_in_advance` in vetKD `ChainKeyConfig` (#5014)
VetKd does not use pre-signatures, and the corresponding config
parameter is ignored. This PR adapts a registry invariant to allow chain
keys that do not require pre-signatures to be configured with zero
pre-signatures to create in advance.

Once this change is rolled out to the mainnet registry, we can update
system tests to configure vetKD keys with zero pre-signatures.