chore: moving farm tests from BN to API BN (#5025)
This PR updates all farm tests that used to rely on boundary nodes for
API calls to use the API boundary nodes.

In addition, it removes a couple of boundary node farm tests that are
now obsolete:
* `bn_integration_test` has been turned into the
`api_bn_integration_test` and the HTTP gateway specific tests are moved
to `dfinity/ic-gateway`.
* `bn_reboot_test` is not needed anymore.
* `custom_domains_integration_test` has been manual for over a year and
hasn't been used in a long time. It is not needed anymore.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>