fix: flaky //rs/tests/consensus:replica_determinism_test (#5060)
The `//rs/tests/consensus:replica_determinism_test` was flaky because
the malicious node could turn malicious and crash the replica just after
the health check but before creating the agent. If this happened
creating the agent would fail with:
```
2025-05-08 10:06:44.846 INFO[test:StdErr] Could not create agent: TransportError(reqwest::Error 
  { kind: Request, url: "http://[2602:fb2b:100:10:501e:e3ff:fefe:d2e5]:8080/api/v2/status", 
    source: hyper_util::client::legacy::Error(Connect, ConnectError(
      "tcp connect error", Os { code: 111, kind: ConnectionRefused, message: "Connection refused" })) })
```
This is fixed by retrying creating the agent because the crashed replica
should come back online after a while.

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>