test(replay): Unit test for Player::get_changes_since in rs/replay. (#5022)
The main thing this demonstrates is reconstituting chunked records,
which is the new functionality being added right now in [PR 4947].

[PR 4947]: https://github.com/dfinity/ic/pull/4947

This is NOT an integration tests; rather, the test heavily relies on
mocking.

This also changes the implementation of `Player::get_changes_since` by
making it call a new stand alone `get_changes_since` function. This
seems effective, because `Player::get_changes_since` does not use any
real data in `Player`. Rather, all it needs is a tokio Runtime, and a
way to make query calls (to the Registry canister).

# References

This verifies production code changes in [PR 4947]

[PR 4947]: https://github.com/dfinity/ic/pull/4947