chore(sns): Publish sns-testing bundle (#4937)
This PR adds a new Bazel target `//rs/sns/testing:sns_testing_bundle`
that bundles all the files required for running sns-testing, including
the binaries and canister wasms. This target's artifact is a single
sns_testing_bundle.tar.gz file that should be published to the CDN (the
size is around 100 MB).