feat(IDX): Add prefix to artifact bundles
This adds a `prefix` option when creating artifact bundles. This will
allow us to use the bundles directly when uploading artifacts to the
CDN, instead of duplicating the logic of listing artifacts & computing
checksums.

This also makes the necessary changes in `build-ic.sh` to support nested
build artifacts as well as fixes a typo in the guestos directory name.

The implementation of `artifact_bundle` is also simplified to avoid
creating symlinks from Bazel; moreover the output of `artifact_bundle`
is now a _directory_ instead of a list of artifacts.

Prefixes are also added to all artifact bundles to reflect the file
structure on the CDN.