perf(idkg): [CON-1497] Performance improvement when building a transcript of a certain ID (#4999)
Currently, when building an IDKG transcript of a certain transcript ID,
we iterate over all dealings and support shares in the artifact pool,
even those that correspond to different transcript IDs, and then filter
out the ones we don’t need. This does not scale under high load.
The artifact pool stores artifacts sorted by their keys. The keys of
dealings and support shares begin with the transcript ID.

This PR thus adds the ability to iterate over all dealings and support
shares of a particular transcript ID, avoiding the overhead described
above.

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>