chore: Don't deliver ALREADY_EXISTS checkpoint in state sync (#5121)
If we try to deliver state synced state, but the execution has already
created the same height checkpoint, we still deliver state sync and load
the execution checkpoint to verify as if it was a state synced
checkpoint.
This change aborts state sync if the destination height is already
reached by execution.