test(nervous_system/canister): Test getting large Registry mutations from a canister. (#5043)
More precisely, the code under test here is `fn registry_changes_since`
(provided by `nervous_system/canisters`). This is meant to be called
from within a canister (and it is not clear to me how you would call it
outside of a canister; calling it outside of a canister might be
impossible).

Therefore, to exercise the code under test, this change introduces a new
test canister: `fetch_large_record_test_canister`, which calls
`registry_changes_since` (i.e. the code under test).

IIUC, the reason `registry_changes_since` cannot be called outside of a
canister is that it (indirectly) uses `ic-cdk` to make canister calls.

# Reference(s)

This verifies production code changes in [PR 4947]

[PR 4947]: https://github.com/dfinity/ic/pull/4947