perf(nns): Use a proper proposal payload in benchmarks (#5110)
# Why

`ProposalData::topic` actually uses 2.5M instructions when topic is
`Unspecified` as it prints the proposal. This distorts the benchmarks,
especially for smaller ones. In a future PR, `ProposalData::topic` will
simply use the topic stored at creation time, which makes the
instructions go away. This PR prepares for that, so that the next PR
doesn't introduce unnecessary benchmark updates which would make the
next PR more difficult to understand.

# What

* Add motion action in proposals added for benchmark preparation
* Update canbench_results.yml