chore: Revert ic-cdk to 0.17.2 (#5139)
This is a full revert of the [bump of ic-cdk to
0.18.0](https://github.com/dfinity/ic/pull/5005). Mixing version 0.18.x
with 0.17.x caused unexpected issues; this is [not possible
anymore](https://github.com/dfinity/cdk-rs/blob/links-pin/TROUBLESHOOTING.md).

With the exception of message routing related testing canisters, which
need a version without this check to test bounded wait call related
code, this means that the monorepo must be upgrade all at once.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>