chore(IDX): Group artifact uploads
This groups all artifact upload jobs into a single target,
`//:upload-artifacts`. This will allow us to turn this into a single
runnable target that can be called separately from the build.

This changes `upload_artifacts` to accept _directories_ instead of
files. The "subdir" used as a bucket path is derived from the relative
path from the directory (budle) to the artifacts themselves. This also
means we can deduplicate the checksumming logic between the
`upload_artifacts` and `artifact_bundle`, since `upload_artifacts` now
only cares about the upload -- not the `sha256sum`.