feat: [MR-680] Use actual XNet streams in `state_manager` tests
Update various `state_manager` tests to encode actual A -> B cross subnet streams with subnet A's State Manager and decode them with subnet B's state manager, instead of taking the simpler path of a loopback stream (encoded and decoded by the same strate manager).

Also have `tree_hash::tests::test_backward_compatibility()` certify a state with both a loopback and a cross-subnet stream.

This is in preparation for dropping the loopback stream from a future version of the certified state.