feat: Write protobufs asynchronously [MR-355] (#4792)
During checkpointing we serialize protobufs of canisters and snapshots
on the critical path, even though we don't need them to continue
execution. In this change we serialize them in the background in
parallel with execution.
Their values are used to validate that the serialized data matches data
in memory (ValidateEq). This step is a background operation.

---------

Co-authored-by: Stefan Schneider <31004026+schneiderstefan@users.noreply.github.com>