test(ICRC_Rosetta): FI-1725: Fix flaky test_deriving_gaps_from_storage test (#5024)
Some flakiness was observed for the `test_deriving_gaps_from_storage`.
This seems to be because the `valid_blockchain_with_gaps_strategy` could
generate some blockchains and corresponding block index values that
actually did not contain gaps. This PR proposes to add some filters to
the strategies, so that the valid blockchain that is generated is of at
least length `2`, and so that the first gap occurs before the last block
in the blockchain.