feat(registry): Make callers of `get_value` understand chunked responses. (#5057)
This mainly involves two things:

1. Making `deserialize_get_value_response` return high-capacity type.
2. Adding `dechunkify_get_value_response_content`, which would usually
be called after `deserialize_*`.

Secondarily, callers of `deserialize_*` also need to call
`dechunkify_*`. This is where most of the changes happen.
    a. nns/common
    b. _(test-only) nns/test_utils_
    c. registry/nns_data_provider
    d. replay

# Alternative Approach

Another way we could have done this was to make `deserialize_` also
dechunkify. However, in an [earlier PR], there was feedback requesting
that a similar `deserialize_*` function not also dechunkify (to wit,
`deserialize_get_changes_since_response`). To be consistent with that,
deserializing and dechunkifying are done separately here as well.

[earlier PR]: https://github.com/dfinity/ic/pull/4947

# References

Closes [NNS1-3685].

[NNS1-3685]: https://dfinity.atlassian.net/browse/NNS1-3685