chore(crypto): CRP-2038 Remove workaround for RSA signature malleability (#5080)
An old version of the `rsa` crate had issues with signature
malleability, where it would accept signatures that were zero-padded to
be longer than the modulus, as well as signatures which were greater
than the public modulus.

These were fixed in a later version of the `rsa` crate, which we have
already upgraded to. But the workaround to check for the malformed
signatures was not removed at that time.