feat(CRP-2702): Log the chain keys on every CUP height (#5150)
This PR adds a log message on every CUP height, that simply prints the
existing public keys of the Subnet as a INFO message.

The Ticket states, that this should only happen for VetKD keys, but I
found it very convenient to simply log all chain keys.
The log message only appears, if there are any chain keys present on the
subnet and only on CUP heights, so the log should not be too spammy