fix: timestamps in nanoseconds in the registry  (#5076)
This PR aligns the naming of timestamps and units being used
(nanoseconds). An overlook was made on the last pr #5023 and the team
decided to continue with nanoseconds because they cause the least amount
of issues.

On top of that the PR addresses the comments left in #5023 that came
after the merge.

Also removes the need for `+ / - 1` in the max delta size calculation.