test(replay): Integration test for `get_changes_since` in replay. (#5164)
This piggy backs off an existing test for a similar function (to wit,
`registry_get_value`).

# Secondary Change(s)

I made `get_changes_since` be `async` instead of taking a `Runtime`
(which was used to `block_on`). This is orthogonal, but it makes
everything a little bit simpler, because tests do not then have to
create a `Runtime`.