test: fix `replica_determinism_test` (#5128)
Removed re-creation of the Agent as it's not necessary (and can fail if
the malicious node is restarting).

Also, modified `await_status_is_healthy` because I was running into
[this reqwest issue](github.com/seanmonstar/reqwest/issues/1233).
[Reqwest's
documentation](https://docs.rs/reqwest/latest/reqwest/blocking/index.html)
says that `Conversely, the functionality in reqwest::blocking must not
be executed within an async runtime, or it will panic when attempting to
block.` which is what we do: i.e. we indirectly use
[reqwest::blocking::Client](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/driver/src/driver/test_env_api.rs?L1570)
within tokio's async runtime
[here](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/consensus/replica_determinism_test.rs?L128-131).