feat: [MR-660] Use actual cross-subnet streams in `state_manager` tests (#5152)
Update various `state_manager` tests to encode actual A -> B cross
subnet streams with subnet A's State Manager and decode them with subnet
B's State Manager, instead of taking the simpler path of a loopback
stream (encoded and decoded by the same State Manager).

Also have `tree_hash::tests::test_backward_compatibility()` certify a
state with both a loopback and a cross-subnet stream.

This is in preparation for no longer including the loopback stream in a
future version of the certified state.