chore: EXC-2033: Improve benchmark scripts (#5160)
This PR adds the following updates/features to the benchmark scripts:

1. By default, all benchmark results are now stored in a file named
`<benchmark>.<host>@<commit>.min`, and the baseline is expected to be in
a `<benchmark>.<host>.min` file.

This helps prevent accidental comparison of results from different hosts
or commit IDs.

Both commit ID and host can be overridden by setting the `COMMIT_ID` and
`FLAVOR` environment variables.

2. The top-level script `run-all-benchmarks.sh` now recognizes the
`INCLUDE` and `FILTER` environment variables.

The `INCLUDE` variable is a globbing pattern used to run only the
benchmark suites matching the pattern.

The `FILTER` variable is a regular expression passed to Criterion to
filter the benchmarks.

3. For better readability of shorter runtimes, all scripts now display
the results in `%.1f ms` format.

4. The `NOISE_THRESHOLD_PCT` and `TOP_N` can now be overridden by
setting the corresponding environment variables.

5. The `summarize-results.sh` script now always produces top N results,
even if there are no overall changes in benchmark results.