test: Remove some usages of install_canister_with_allocation (#5171)
This PR refactors a few tests that are using the
`install_canister_with_allocation` function to manipulate allocations in
an attempt to check certain conditions. Given that the path of updating
the allocations via `install_code` is considered deprecated, the tests
are transformed to avoid the use of it. Mainly this is achieved via
setting the required allocations at creation time, with one exception
where the test is changed to use another case of failure (instead of
relying on passing an invalid allocation).