refactor(cketh): ensure the EthRpcClient always gets an evm_rpc_id (#5055)
[XC-224](https://dfinity.atlassian.net/browse/XC-224?atlOrigin=eyJpIjoiOTdjOGU5OWFlNzMwNGYxMWJiYzhhZmM0Y2NlZGQxODgiLCJwIjoiaiJ9):
This PR continues the effort to remove legacy code that enabled the
ckETH minter to interact with Ethereum via direct HTTPS outcalls (see
[previous PR](https://github.com/dfinity/ic/pull/4926)).

This PR preserves backward compatibility by keeping the `evm_rpc_id`
field optional in the upgrade args and adding it as optional in the init
args. This PRs ensures that there is always an `evm_rpc_id` in the state
read by the ETH RPC client. The `evm_rpc_id` defaults to the production
(mainnet) or staging (sepolia) EVM RPC ID when it is left unspecified in
the init and upgrade args.


[XC-224]:
https://dfinity.atlassian.net/browse/XC-224?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ