test(registry): More tests for large records/chunking. (#5178)
This addresses all `TODO(NNS1-3746)`s by either adding more tests, or
enhancing existing ones, now that we have chunking.

All tests in

* `rs/registry`
* `rs/nns`

ALSO pass when I enable chunking (by setting
`IS_CHUNKIFYING_LARGE_VALUES_ENABLED` to true), so we should be able to
simply flip the flag when the time comes.

# Very Slightly More Chunking

Since chunking is still disabled in release builds, there is
technically, no behavior change here, and so `unreleased_changelog.md`
does not need to be updated. However, if chunking WERE enabled, there
would be a (small) behavior change: the min chunkable size has been
decreased by around 100 bytes.

We never promised that the previous value is EXACTLY EXACTLY EXACTLY
when chunking would start happening. Rather, we just tried to avoid
chunking as much as possible. With a slightly lower limit, we are still
achieving the goal where we "avoid chunking as much as possible". This
is because 100 bytes is an extremely small amount compared to the new
value, which is (still) around 1_300_000 bytes.

# References

Closes [NNS1-3746].

[NNS1-3746]: https://dfinity.atlassian.net/browse/NNS1-3746