fix: don't run the head_nns variants of the nested tests during system_test_nightly (#5218)
The nested system-tests are currently failing due to
[this](https://github.com/dfinity/ic/pull/5089#issuecomment-2892408601).
The problem is that the `head_nns` variants of the nested system-tests
run during the "Bazel System Test Nightly" job which is part if "Release
Testing" which means we can't make a new release.

So we make sure the `head_nns` variants don't run on "Bazel System Test
Nightly" but on "Schedule Hourly" instead.