test: move //rs/tests/nested:registration from hourly to long_test (#5231)
This moves the `//rs/tests/nested:registration` test from the "Schedule
Hourly" workflow to run on every push to master (`long_test`). This has
the advantage that bugs are discovered sooner and can be pinpointed more
quickly to the commit that caused it.

The intention is that all `//rs/tests/nested:...` tests are moved to
`long_test`. However, the `//rs/tests/nested:hostos_upgrade...` tests
are currently failing because of
[this](https://github.com/dfinity/ic/pull/5089#issuecomment-2892408601).