chore: Remove code handling compute and memory allocations in install_code (#5204)
This PR removes all code that is handling updating compute and memory
allocations via `install_code` which is a deprecated path. The fields
still exist in `InstallCodeArgs` but they are essentially ignored.
Alongside the code a bunch of tests are removed that were checking for
the conditions handled by the removed code.

In a follow-up PR the fields will be completely dropped from
`InstallCodeArgs` -- this would allow that PR to be fully mechanical and
easy to review (also it would be touching code of other teams as well).