fix: Use correct arguments for stable_read in WatCanisterBuilder (#5226)
The arguments used for `stable_read` in WatCanisterBuilder were
incorrect, basically missing the `dst` argument. This meant that
basically `stable_read` really depended on a `stable_grow` to exist
before it so that it can use its return value as one of the arguments.
While this may be acceptable (how can you read before growing your
memory first?), it also creates the reverse dependency. If one uses a
`stable_grow` only, it will leave one operand on the stack and the
resulting wasm module would be invalid.

One fix (which is what this PR proposes) is to always drop the result of
`stable_grow` -- in a test utility it's not particularly useful anyway,
most of the time the test would expect to succeed and even if it wants
to fail, this can (and should be) tested outside the Wasm module (e.g.
by inspecting the canister's state in the test. This means that
`stable_read` would also need to be fixed to accept 3 arguments as per
the [interface
spec](https://internetcomputer.org/docs/references/ic-interface-spec#system-api-stable-memory).