fix: canister http response size validation (#5234)
This PR implements the following changes to the canister http response
size validation:
- (i) the transformed response size must not exceed `max_response_bytes`
(if provided, otherwise the default value is used just like now);
- (ii) drops the check for the _serialized_ response size if no
transform is requested.

The first change (i) is a potentially breaking change: logs have been
introduced by this [PR](https://github.com/dfinity/ic/pull/4873) rolled
out to all subnets by May 5, 2025, with no hits so far.

The reasoning behind the second change (ii) is that the response size
(as defined in the [Interface
Specification](https://internetcomputer.org/docs/references/ic-interface-spec#ic-http_request),
i.e., "the total number of bytes representing the names and values of
HTTP headers and the HTTP body") is already
[checked](https://github.com/dfinity/ic/blob/d3d29e47f32c834dedf8abbc187f4557ab76bef7/rs/https_outcalls/adapter/src/rpc_server.rs#L528-L542)
by the adapter and thus repeating the check after serialization is not
necessary (since the serialization overhead is small and constant) and
could be confusing (since the check is not specified to be performed).

This PR also ensure that `transform_timer.observe_duration()` is always
observed (also in error cases).