refactor: Load entire kernel cmdline from boot partition (#5101)
This PR introduces a new mechanism to handle kernel boot arguments
across GuestOS, HostOS and SetupOS.

Previously, some arguments were calculated in grub.cfg while others were
loaded from the `extra_boot_args` file from the boot partition. Then
`grub.cfg` concatenated the arguments.

This current method had a few shortcomings:
- The arguments coming from grub.cfg were never updated because the
grub.cfg is not changed during OS upgrades.
- SEV-SNP measurements depend on the kernel cmdline, so it's critical to
have the exact cmdline which will be tied to the GuestOS release. This
is attainable only if we load the entire cmdline from the boot
partition.

We introduce a new file on the boot partition: `boot_args`. This is
parallel to the existing `extra_boot_args`. In case of GuestOS and
HostOS, we store two variables `BOOT_ARGS_A` and `BOOT_ARGS_B` which
correspond to the cmdline when launching as boot alternative A and B. In
the SetupOS, we only store a single variable `BOOT_ARGS`.

To allow for some flexibility when upgrading/downgrading, we continue to
support the previous boot mechanism if the `boot_args` file is not
found.

---------

Co-authored-by: r-birkner <103420898+r-birkner@users.noreply.github.com>