fix(icrc-rosetta): Fix timestamp overflow (#5249)
In Rosetta ICRC's `blocks` table, the timestamp fields `timestamp`,
`transaction_created_at_time` and `approval_expires_at` are defined with
`INTEGER` type, which is a *signed* 64-bit integer.

However, in the ledger's internal representation those are *unsigned*
64-bit integers, so when the value is over `i64::MAX`, it fails to store
it in the database. Since `approval_expires_at` is chosen by the user,
they can (and have) used such values, which makes the synchronization
fail entirely.

Although the other two fields are automatically defined, we're fixing
them too for the sake of consistency and fool-proofing.

This problem doesn't happen with ICP Rosetta because it stores
timestamps in `TEXT` fields, which we should replicate eventually in
ICRC1 Rosetta, although it's a much more disruptive change.