chore(PocketIC): set time on the server for auto-progress mode (#5262)
This PR refactors PocketIC to set the (certified) time for auto-progress
mode on the server instead of in the library: this way, trying to enable
the auto-progress mode on a PocketIC instance with the auto-progress
mode already enabled is a no-op (instead of resetting the certified time
again).

This PR also fixes doc comments to not distinguish between `/api/v2` and
`/api/v3` requests since more versions might be introduced in the future
and the doc comments are already inconsistent (some only refer to
`/api/v2`).