refactor(ckbtc): Simplify ckbtc minter tests (#5169)
While reading the ckbtc tests, I noticed that there is some code to
enable the ECDSA signatures on the subnet, that is no longer needed,
since the subnet can simply be enabled with the corresponding ECDSA key
from the start.

This means that some code to enable the key can be removed, simplifying
the test.

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>