refactor(CON-1446): Move DkgPayloadBuidler error types into ic-types (#4960)
This is a PR in preparation of creating a `DkgPayloadBuilder` trait and
a `DkgPayloadBuilderImpl` struct.

To define the trait, the `ic-interfaces` trait needs to have access to
the error types. Therefore, we move the types from the
`ic-consensus-dkg` crate into the `ic-types` crate, such that they can
references in the `ic-interfaces` crate, to define the
`DkgPayloadBuilder` trait.

Additionally, it renames some of the types by prefixing them with `Dkg`,
such that it is less likely to confuse them with their IDkg or top-level
counterparts, such as:

- `Summary` -> `DkgSummary`
- `Payload` -> `DkgPayload`
- `PayloadValidationError` -> `DkgPayloadValidationError`

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>