chore(drun): eliminate backtraces when testing with `drun` (#5272)
Since `moc` tests are run in a controlled environment, we do exactly
know which traps occur and when. So there is zero added value in
observing backtraces, but they tend to diverge in various test targets
(e.g. EOP vs. classical) so to avoid normalisation for golden tests,
better suppress them.