chore: Remove compute and memory allocation from `InstallCodeArgs` (#5190)
This PR removes the deprecated `compute_allocation` and
`memory_allocation` fields from `InstallCodeArgs` (and
`InstallCodeArgsV2`). The change is safe to perform as the fields are
not used according to internal metrics and they are optional in the
candid spec.

The changes included are mostly mechanical and follow from the main goal
of removing the fields:
1. Printing the fields is removed from the `Display` of
`InstallCodeArgs` and `InstallCodeArgsV2`.
2. A lot of call sites (mostly tests) are updated to not have the fields
anymore.
3. The respective fields are also dropped from `InstallCodeContext`
(along with the code parsing them).