test(nns): Stop assuming voting with well known neurons must succeed as they might follow each other (#5280)
When simulating voting with multiple neurons in the test, we should not
assume the voting all succeed since they might follow each other. On the
other hand, not checking the voting result is OK since we will wait for
the proposal execution anyway. The only downside of not checking is that
when proposal does not execute, it's slightly more difficult to debug.