fix(sns): Avoid calling NNS Governance during swap finalization if Neuron's Fund participation was not requested (#5283)
This PR fixes an issue that could occur if the NNS Governance canister
does not find proposals that correspond to the creation of the SNS that
requests Neurons' Fund participation during finalization. Note that
currently, the Neurons' Fund is disabled, and the SNSs that do not even
request Neurons' Fund participation are potentially risking that their
finalization halts if the NNS proposal that created that SNS cannot be
found (this recently happened due to an unrelated problem in the NNS).

The solution is to avoid calling the NNS Governance's
`settle_neurons_fund_participation_result` function is the SNS des not
specifically request Neurons' Fund participation.