refactor: Add new generate-guestos-vm-config feature to ic-os config tool which replaces generate-guestos-config.sh (#5264)
A few motivators of the new implementation:

- Typesafe implementation.
- The XML generation now happens in Rust using Jinja-like templates
which are more powerful than sed (e.g. loops, conditionals are
supported) and are checked at compile-time. The templates are easier to
read.
- Tests were added including golden files for the XML output.
- Previously we had two shell scripts (generate-guestos-config.sh and
dev-generate-guestos-config.sh) which only differed in a few lines and
had to be otherwise kept in sync. This is no longer needed, since we can
use rust feature flags to enable dev behavior conditionally.
- Bonus: it works faster (0.7s -> 0.03s)

I manually verified that the behavior of the new tool is the same as the
old bash script. The only difference is that we create the guestos
config in a temp file instead of `/boot/config/config-guestos.json`.
Since this file is only used to create the bootstrap config image, it's
not necessary to store it on the config partition. In a later PR, we can
completely remove the temporary file and just pass around a Rust struct.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: Andrew Battat <113942931+andrewbattat@users.noreply.github.com>
Co-authored-by: Bas van Dijk <bas@dfinity.org>