refactor(ic-boundary): Remove dependency to canister_client in ic-boundary (#5288)
The `canister_client` crate is not very maintained anymore and we would
like to avoid its use. This PR replaces `canister_client` with
`ic-agent` in the `ic-boundary` crate.

`canister_client` and `ic-agent` have similar APIs, so only a few
adjustements to the inputs were necessary to match the required
types/syntax:
- A new `NodeSender` struct is introduced to mirror the functionality of
[`Sender::Node`](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/canister_client/sender/src/lib.rs?L118),
which is in `canister_client` and that was used here.
- Responses are `Option<Vec<u8>>` in `canister_client` but `Vec<u8>` in
`ic-agent`. The case where the response is `None` in the former is a 1-1
relationship to having an empty `Vec` in the latter (it is actually
impossible to have a `Some(vec![])` in the former). This justifies the
changes handling the responses.
- Nonces are generated directly inside `ic-agent` and the caller does
not have this responsibility anymore.

---------

Co-authored-by: Igor Novgorodov <igor.novgorodov@dfinity.org>
Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>