fix(icp-rosetta): fix heartbeat during initial sync (#5293)
This pull request introduces a new mechanism to ensure that the
`heartbeat_fn` is only invoked after the first successful blockchain
synchronization in the `start_sync_thread` function. This change
improves the logic for handling synchronization and ensures that the
heartbeat is not prematurely triggered.

Without this, if there's any failure during the initial sync, the
watchdog thread will think the initial sync is done and only wait up to
60 seconds per cycle, leading to multiple sync thread restarts.