chore: Remove defrag_canisters_map step (#5313)
The defrag_canisters_map step during checkpointing was initially
intended to be a functional no-op, while improving performance. In
particular, it was intended to combat the slight performance difference
between a subnet that was recently upgraded, and one that has been
running for a while. However, it turned out that the step never lead to
the desired significant improvements.

Now that checkpointing has been optimized to the area of 1 second with
asynchronous checkpointing, the defrag_canisters_map step contributes as
much as 40% to the checkpointing time on some subnets, which makes it
not worth it.

This PR removes the step without replacement.