fix(rosetta-icrc): fee collector balance fix (#5304)
The balance calculation wasn't taking into account fees being sent to
fee collectors defined in a different block, which led to fee collector
accounts having wrong balances in the local Rosetta DB.

This PR:
- Makes sure that fee collectors' balances are correctly modified in
such cases
- Reconciles balances of DBs potentially affected by the bug at startup
by recreating the balances table. We keep track when this fix has
happened for a particular DB so we only do it once.

Tested it manually with GLDT token, which was crashing due to this
issue:
- Backfilled with broken revision until crash. Restarted with the fix,
balances were conciliated and operation recovered. Restarting again
didn't cause a new reconciliation.
- Backfilled with the fixed revision. Restarting it didn't cause a
reconciliation.