feat(icrc-rosetta): add icrc rosetta release 1.2.3
Release information for ICRC Rosetta 1.2.3 with the following changes:

- Fixed fee collector balance calculation for transfers using fee_collector_block_index

The balance calculation wasn't taking into account fees being sent to
fee collectors defined in a different block, which led to fee collector
accounts having wrong balances in the local Rosetta DB.

This release:
- Makes sure that fee collectors' balances are correctly modified in
such cases
- Reconciles balances of DBs potentially affected by the bug at startup
by recreating the balances table. We keep track when this fix has
happened for a particular DB so we only do it once.