fix: full suffix for sha matching in tests (#5323)
We had an incident (fixed in #5305) where all of the artifacts were
uploaded to the same SHA256SUMS file on our object storages. This causes
the current test logic to assume and expect wrong SHA since the SHAs are
sorted in an ascending order in the SHA256SUMS.

In this logic we only download update images and this is just a stronger
requirement that will prevent the explained bug from happening again.
Currently this breaks our qualification system test since the node is
constantly expecting a wrong SHA.

An example file that breaks the code and is fixed by this pr can be
downloaded
[here](http://download.proxy-global.dfinity.network:8080/ic/181968dd5f36af93590490802540caf7d2df915a/guest-os/update-img/SHA256SUMS)