docs(nns): Get rid of instructions for obsolete way of doing upgrade testing. (#5312)
The old way was hyper flaky and super slow. We have not used that in a
long time, because the new way seems give us the same (if not more)
assurance, is much more reliable, and much faster.

The new way uses `StateMachine`'s ability to take a `state_dir`.

The old way used `recovered_mainnet_nns`, which is [being deleted].

[being deleted]: https://github.com/dfinity/ic/pull/5307