test(nns): add maturity disbursement tla model and code link (#5225)
Maturity disbursement has been recently added to the NNS code base. This
updates the TLA model of governance to also model these functions, and
adds the instrumentation to link the model to the code.

This is a bigger change since the new code violated some assumptions the
existing TLA instrumentation made about the code structure. In
particular, so far, all the entry points were methods on `Governance`,
where's the new timer is a free-standing function that takes a
`LocalKey` to the governance, and we didn't have to deal with so many
`async_trait`s. I overhauled the instrumentation to make it more
flexible and also cleaned some of the code up.

I also added some usage documentation for the instrumentation.