feat(nns): Expose a metric voting_power_snapshots_latest_snapshot_is_spike (#5179)
A voting power spike is a significant signal, so we'd like to expose the
metric allowing monitoring system to pick up the signal.

A benchmark `voting_power_snapshots_is_latest_snapshot_a_spike` is added
as the voting power snapshots can be quite big, and the fact that we can
call `is_latest_snapshot_a_spike` in metrics endpoint is due to the fact
that the per-neuron voting power is stored in a different StableBTreeMap
than the totals, and the per-neuron data is not deserialized during
is_latest_snapshot_a_spike. The benchmark should prevent future
modifications from incorrectly use per neuron data within
`is_latest_snapshot_a_spike`.